Create extension to paged query
